Updating the main contents.
The main contents was just supposed to be the tutorial first and then links to the three sets of tool documentation. The "temporary" links to the tutorials just got added as place-holders. Now the MAP Client docs are included, we can remove the temp link. Closes #7, but perhaps some more work needs to be done.